pkg-py/tests/playwright/chat/pydantic-ai/requirements.txt
chatlas
faicons
openai
pydantic-ai
python-dotenv
shiny
